refactor: restructure templates
