+++ src/templates/component/componentJsTemplate.js
