+++ src/templates/component/componentLazyTemplate.js
