+++ src/templates/component/componentStoryTemplate.js
