+++ src/templates/component/componentTestDefaultTemplate.js
