+++ src/templates/component/componentTestEnzymeTemplate.js
