+++ src/templates/component/componentTestTestingLibraryTemplate.js
